Reproject EASE-Grid 2.0 SMAP data to Web Mercator

+++ autodownload_smap_36km.py
@@ -4,17 +4,21 @@ import os
 import geopandas as gpd
 import mercantile
 import rasterio
+import affine
 from rasterio.crs import CRS
-from rasterio.warp import reproject, calculate_default_transform
+from rasterio.warp import Resampling, reproject, calculate_default_transform
 import h5py
 from datetime import datetime, timedelta
 import requests
 
-username = os.environ["EARTHDATA_USER"] #"YOUR-USERNAME"
-password = input("Enter EarthData password: ") #"YOUR-PASSWORD"
+username = os.environ["EARTHDATA_USER"]
+password = os.environ["EARTHDATA_PASS"]
 
 os.chdir(os.path.dirname(__file__))
 
+ZOOM = 3
+TILE_SIZE = 128
+RESAMPLING = Resampling.bilinear
 # Output stored under ROOT_PATH/raw/
 #ROOT_PATH = "/s/" + socket.gethostname() + "/b/nobackup/galileo/sm_predictions/daily_predictions/input_datasets/smap_36"
 ROOT_PATH = "data"
@@ -29,42 +33,59 @@ def download_smap_automatically(year=None, month=None, day=None, n_days_before=2
         current_date = datetime.now()
         year, month, day = (current_date - timedelta(days=n_days_before)).strftime("%Y-%m-%d").split("-")
 
-    host = 'https://n5eil01u.ecs.nsidc.org'
-    version = '009'  # product version
-    url_path = '{}/SMAP/SPL3SMP.{}/{}.{}.{}/'.format(host, version, year, month, day)
-
-    file_version = '001'
-    filename = 'SMAP_L3_SM_P_{}{}{}_R19240_{}.h5'.format(year, month, day, file_version)
-    smap_data_path = url_path + filename
     out_dir = ROOT_PATH + "/raw/"
     os.makedirs(out_dir, exist_ok=True)
 
+    filename_v1 = get_filename(year, month, day, '001')
+    filename_v2 = get_filename(year, month, day, '002')
+
+    path_v1 = os.path.join(out_dir, filename_v1)
+    path_v2 = os.path.join(out_dir, filename_v2)
+
+    if os.path.isfile(path_v1) or os.path.isfile(path_v2):
+        return
+
     with requests.Session() as session:
         session.auth = (username, password)
-        filepath = os.path.join(out_dir, filename)
-        response = session.get(smap_data_path)
+
+        filename = filename_v1
+        filepath = path_v1
+        response = retrieve_file(session, year, month, day, filename_v1)
 
         if not response.ok:
-            print('Error downloading with file_version="001". Retrying with file_version="002"')
+            print(f'Error downloading with file_version="001". Retrying with file_version="002"')
+            filename = filename_v2
+            filepath = path_v2
+            response = retrieve_file(session, year, month, day, filename_v2)
 
-            file_version = '002'
-            filename = 'SMAP_L3_SM_P_{}{}{}_R19240_{}.h5'.format(year, month, day, file_version)
-            smap_data_path = url_path + filename
-            filepath = os.path.join(out_dir, filename)
-            print("PATHL ", smap_data_path)
-
-            response = session.get(smap_data_path)
-
-        if response.status_code == 401:
-            response = session.get(response.url)
         assert response.ok, 'Problem downloading data! Reason: {}'.format(response.reason)
-
+    
         with open(filepath, 'wb') as f:
             f.write(response.content)
 
         print(filename + ' downloaded')
         print('Downloading SMAP data for: ' + str(year) + '-' + str(month).zfill(2) + '-' + str(day).zfill(2))
-    #load_file_h5()
+
+def get_filename(year, month, day, file_version):
+    return 'SMAP_L3_SM_P_{}{}{}_R19240_{}.h5'.format(year, month, day, file_version)
+
+def retrieve_file(session, year, month, day, filename):
+    host = 'https://n5eil01u.ecs.nsidc.org'
+    version = '009'  # product version
+    url_path = '{}/SMAP/SPL3SMP.{}/{}.{}.{}/'.format(host, version, year, month, day)
+    
+    smap_data_path = url_path + filename
+
+    print("PATHL ", smap_data_path)
+
+    response = session.get(smap_data_path)
+
+    if response.status_code == 401:
+        # why have you done this to us, HTTP
+        response = session.get(response.url)
+
+    return response
+
 
 def list_all_bands_in_h5_file(file_path):
     def list_datasets(name, obj):
@@ -79,18 +100,37 @@ def list_all_bands_in_h5_file(file_path):
 def create_geotiff(data, output_file):
     src_crs = CRS.from_epsg(6933) # EASE-Grid 2.0 Global
     dst_crs = CRS.from_epsg(3857) # WGS 84 / Pseudo-Mercator
-    transform = calculate_default_transform(
-        src_crs, dst_crs, data.shape[1], data.shape[0], 
+
+    # https://nsidc.org/data/user-resources/help-center/guide-ease-grids#anchor-36-km-resolution-ease-grids
+    x_topleft = -17367530.44
+    y_topleft = 7314540.83
+    grid_units = 36032.22
+
+    x_bottomright = x_topleft + grid_units * data.shape[1]
+    y_bottomright = y_topleft - grid_units * data.shape[0]
+
+    src_transform = affine.Affine.from_gdal(x_topleft, grid_units, 0, y_topleft, 0, -grid_units)
+
+    transform, width, height = calculate_default_transform(
+        src_crs, dst_crs, data.shape[1], data.shape[0],
+        left=x_topleft, bottom=y_bottomright, right=x_bottomright, top=y_topleft
     )
-    with rasterio.open(output_file, 'w', driver='GTiff', width=data.shape[1], height=data.shape[0], count=1, crs=crs, ) as dataset:
-        #driver = gdal.GetDriverByName('GTiff')
-        #dataset = driver.Create(output_file, data.shape[1], data.shape[0], 1, gdal.GDT_Float32)
-        #dataset.GetRasterBand(1).WriteArray(data)
-        #geotransform = (-180.00, 0.373443973830433, 0, 85.0445, 0, -0.373443973830433)
-        #dataset.SetGeoTransform(geotransform)
-        #dataset.SetProjection("EPSG:4326")
-        #dataset = None
-        dataset.write(data, 1)
+
+    nodata = -9999.0
+
+    with rasterio.open(
+        output_file, 'w', driver='GTiff',
+        width=width, height=height, count=1, crs=dst_crs, 
+        transform=transform, dtype=data.dtype, nodata=nodata
+    ) as dataset:
+        reproject(
+            source=data, destination=rasterio.band(dataset, 1),
+            src_transform=src_transform, src_crs=src_crs,
+            dst_transform=transform,
+            dst_crs=dst_crs,
+            resampling=RESAMPLING,
+            src_nodata=nodata
+        )
 
 
 def load_file_h5():
@@ -126,13 +166,13 @@ def load_file_h5():
 
 
 def chop_in_quadhash():
-    quadhash_dir = next(d for d in os.listdir() if os.path.isdir(d) and d.startswith("quadshape_3_"))
+    #quadhash_dir = next(d for d in os.listdir() if os.path.isdir(d) and d.startswith(f"quadshape_{ZOOM}_"))
     #quadhashes = gpd.read_file(os.path.join(quadhash_dir, 'quadhash.shp'))
     #tiles = mercantile.tiles(-180, -85, 180, 85, [3], True)
-    tiles = mercantile.tiles(-125, 24, -66, 50, [3], True)
+    tiles = mercantile.tiles(-125, 24, -66, 50, [ZOOM], True)
 
     in_path = ROOT_PATH + "/raw/"
-    out_path = ROOT_PATH + "/split_3/"
+    out_path = ROOT_PATH + f"/split_{ZOOM}/"
     os.makedirs(out_path, exist_ok=True)
 
     count = 0
@@ -169,4 +209,6 @@ if __name__ == '__main__':
     # Remove for loop below if want one day prediction
     for i in range(2,3): #(2,50)
         download_smap_automatically(year=None, month='04', day='04', n_days_before=i)
+    load_file_h5()
+
     #chop_in_quadhash()
